pyproject.toml
[project]
name = "autogit"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.22.0",
    "fastapi>=0.127.0",
    "pydantic-settings>=2.12.0",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.40.0",
]
